renamed the useResponsiveComponent to makeResponsive  and moved to utils + added proper folder structure
