+++ packages/atomic-layout/src/index.ts
@@ -13,8 +13,8 @@ export { default as useViewportChange } from './hooks/useViewportChange'
 export { default as useBreakpointChange } from './hooks/useBreakpointChange'
 export { default as useResponsiveValue } from './hooks/useResponsiveValue'
 export { default as useResponsiveProps } from './hooks/useResponsiveProps'
-export { default as useResponsiveComponent } from './hooks/useResponsiveComponent'
 export { default as useResponsiveQuery } from './hooks/useResponsiveQuery'
 
 /* Utils */
 export { query } from './utils/query'
+export { makeResponsive } from './utils/makeResponsive'
